move the stingy runtime to a common area.
