+++ tests/test_s3copyclient.py
@@ -16,26 +16,7 @@ from dss.blobstore.s3 import S3BlobStore
 from dss.events import chunkedtask
 from dss.events.chunkedtask.s3copyclient import S3CopyTask
 from tests import infra
-
-
-class TestStingyRuntime(chunkedtask.Runtime[dict, bool]):
-    """This is runtime that returns a pre-determined sequence, and then 0s for the remaining time."""
-    def __init__(self, seq=None):
-        self.complete = False
-        if seq is None:
-            seq = list()
-        self.seq = itertools.chain(seq, itertools.repeat(0))
-
-    def get_remaining_time_in_millis(self) -> int:
-        return self.seq.__next__()
-
-    def reschedule_work(self, state: dict):
-        # it's illegal for there to be no state.
-        assert state is not None
-        self.rescheduled_state = state
-
-    def work_complete_callback(self, bool):
-        self.complete = True
+from tests.chunked_worker import TestStingyRuntime
 
 
 class TestAWSCopy(unittest.TestCase):
